+++ TGEngine/drawlib/Circle.hpp
@@ -3,4 +3,4 @@
 #include "BaseDrawLib.hpp"
 
 SINCE(0, 0, 4)
-void drawCircle(TGVertex pos, float radius, int max_verticies, VertexBuffer* vert, IndexBuffer* ibuffer);
+void drawCircle(TGVertex pos, float radius, size_t max_verticies, VertexBuffer* vert, IndexBuffer* ibuffer);
