+++ TGEngine/drawlib/Ellipse.hpp
@@ -3,4 +3,4 @@
 #include "BaseDrawLib.hpp"
 
 SINCE(0, 0, 2)
-void drawEllipse(TGVertex pos, float xradius, float yradius, int max_verticies, VertexBuffer* vert, IndexBuffer* ibuffer);
+void drawEllipse(TGVertex pos, float xradius, float yradius, size_t max_verticies, VertexBuffer* vert, IndexBuffer* ibuffer);
